+++ src/constants/common.js
@@ -131,4 +131,5 @@ export const urlPath = {
 
   //금융아티클
   FINANCIAL: "/financial",
+  MYPAGE: "/mypage"
 };
